requirements.txt
requests
pandas

# database
# sqlite
sqlalchemy

# environment
python-dotenv
